+++ src/index.js
@@ -1,12 +1,17 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './App';
+import FaceRecog from './FaceRecog';
+import TextRecog from './TextRecog';
 import reportWebVitals from './reportWebVitals';
+import { BrowserRouter, Routes, Route} from "react-router-dom";
 
 ReactDOM.render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>,
+  <BrowserRouter>
+    <Routes>
+      <Route path="/" element={<FaceRecog/>}/>
+      <Route path="textRecog" element={<TextRecog/>}/>
+    </Routes>
+  </BrowserRouter>,
   document.getElementById('root')
 );
